api requester update
creating a 'module' file for functions
and a 'usage' file for examples
